pyproject.toml
[tool.poetry]
name = "python-api-rds-postgres"
version = "0.1.0"
description = ""
authors = ["jeas560 <jeas560@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pandas = "^2.1.4"
sqlalchemy = "^2.0.23"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
